moved showAlert into its own file and added configurable title
